mcp/client/requirements.txt
fastapi
fastmcp
python-dotenv==1.2.2
uvicorn
